ship-statics/requirements.txt
# ══════════════════════════════════════════════════════════
# 船舶静力学课程设计自动计算系统 - Python 依赖
# 武汉理工大学 船舶与海洋工程学院
# ══════════════════════════════════════════════════════════

# ── Web 框架 ──
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# ── 数值计算 ──
numpy==1.24.3
scipy==1.11.4

# ── 数据处理 ──
pandas==2.1.3
openpyxl==3.1.2

# ── 图表绘制 ──
matplotlib==3.8.2

# ── Word 文档生成 ──
python-docx==1.1.0

# ── 其他工具 ──
Pillow==10.1.0


